feat: implement chunk and pad util in predicate root calc
